requirements.txt
fastapi
uvicorn

openai
langchain
langchain-openai
langchain-community
langgraph
langsmith
langchain-cohere
langgraph-checkpoint-sqlite

pydantic>=2.7.4,<3
pandas
requests

SQLAlchemy
aiosqlite
loguru
python-dotenv
psycopg2-binary
aiohttp>=3.9.0

